requirements.txt
django
fastapi
uvicorn
tensorflow
python-multipart
pillow
numpy
requests